requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==39.0.1
dj-database-url==1.2.0
Django==4.1.6
django-adminlte-3==0.1.6
django-adminlte3==0.1.6
django-bootstrap4==22.3
django-localflavor==3.1
djangorestframework==3.14.0
idna==3.4
mysqlclient==2.1.1
Pillow==9.5.0
pycparser==2.21
PyMySQL==1.0.2
python-decouple==3.7
python-dotenv==0.21.1
python-stdnum==1.18
pytz==2022.7.1
requests==2.28.2
soupsieve==2.4
sqlparse==0.4.3
Unipath==1.1
urllib3==1.26.15
